requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==24.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.4
certifi==2024.12.14
chardet==4.0.0
charset-normalizer==3.4.0
click==8.0.3
cloud-init==24.3.1
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
httplib2==0.20.2
huggingface-hub==0.26.5
hyperlink==21.0.0
idna==3.10
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
Mako==1.3.8
MarkupSafe==3.0.2
more-itertools==8.10.0
multidict==6.1.0
netifaces==0.11.0
numpy==2.2.0
oauthlib==3.2.0
packaging==24.2
propcache==0.2.1
psycopg2-binary==2.9.10
py-cord==2.6.1
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycurl==7.44.1
pydantic==2.10.3
pydantic_core==2.27.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dotenv==1.0.1
pytz==2022.1
PyYAML==5.4.1
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
SQLAlchemy==2.0.36
systemd-python==234
tokenizers==0.21.0
tqdm==4.67.1
transformers==4.47.0
Twisted==22.1.0
typing_extensions==4.12.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
yarl==1.18.3
zipp==1.0.0
zope.interface==5.4.0
